+++ src/pages/content/Providers.tsx
@@ -2,12 +2,12 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import "@src/assets/styles/studres-css-reset.css"
 import "@src/assets/styles/shadcn-ui.css"
 import { CommandProvider } from "@src/components/command/CommandContext"
-import ConfigContextProvider from "@src/contexts/ConfigContext"
 import { TooltipProvider } from "@src/components/ui/tooltip"
 import { SmoothRouterListener } from "@src/features/router/SmoothRouterListener"
 import { PreferredTheme, ThemeProvider } from "@src/contexts/ThemeContext"
 import { Toaster } from "@src/components/ui/toaster"
 import { PageStateContextProvider } from "@src/features/router/PageStateContext"
+import ConfigContextProvider from "@src/features/config/ConfigContext"
 
 interface ProvidersProps {
     children: React.ReactNode
